Update patch set 3
Patch Set 3:

Ah. Wasn't aware of the inheritance. 

If we were to create fastboot_unlocked and move A6020 over to that, that would be the best option, and leave fastboot_generic alone, and just make sure we catch any new device page submissions using fastboot_generic, and block those.

Patch-set: 3